Use /var/tmp for large temporary storage
The /tmp file system on our ALP VM images is a tmpfs, meaning that
it's capacity is limited to half the system memory size, which can
be problematic for staging large objects, such as saved container
images. Switching to /var/tmp for this case avoids issues.

Also remove the 'root-' prefix from the Ansible for ALP workload
labels to match recent label naming changes.