Cleanup ansible-lint reported complaints.
Also cleanup the .ansible-lint file to remove rules that no longer
need to be converted to warnings.

Added .ansible-lint-ignore with entries to ignore failures for
commented out lines in YAML settings files under settings/.

Added a Contributing section to the README.md discussing the fact
that the code should pass ansible-lint testing.